Chapter3/Framework_pipeline/pipeline/5_serving/Docker/requirements.txt
seldon-core
numpy
kfserving==0.3.0
cachetools==4.0.0
certifi==2019.11.28
oauthlib==3.1.0
protobuf==3.11.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
tornado==6.0.3
urllib3==1.25.8
websocket-client==0.57.0
tensorflow==2.3.0
imutils
pillow
opencv-contrib-python-headless==4.2.0.34
opencv-python==4.2.0.34